REF: move every policy class into simplesat.sat.policy.
